system-project-management/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
email-validator
python-dotenv
passlib
python-jose[cryptography]
pytest
httpx
